backend/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
pydantic
PyJWT
python-dotenv
python-jose
